Move WithScale to layers package
