+++ src/layers/TextLayer.tsx
@@ -14,8 +14,8 @@ import ThemeContent from '../theme/ThemeContext';
 import calculateOffset from '../utils/calculateOffset';
 import unwrap from '../utils/unwrap';
 import wrap from '../utils/wrap';
-import WithScale from '../WithScale';
 import './textLayer.less';
+import WithScale from './WithScale';
 
 interface TextLayerProps {
     keywordRegexp: RegExp;
